7/15 감정 사전 연결 완료

+++ Spring/src/main/java/com/haeildairy/www/emotion/service/AllService.java
@@ -81,7 +81,7 @@ public class AllService {
      */
     private List<FlaskResponseDTO.FlaskEmotionDetailDTO> convertMoodDetailToFlaskDetail(List<MoodDetailDTO> moodDetailDTOs) {
         return moodDetailDTOs.stream()
-                .map(dto -> new FlaskResponseDTO.FlaskEmotionDetailDTO(dto.getEmotionType(), dto.getPercentage().doubleValue()))
+                .map(dto -> new FlaskResponseDTO.FlaskEmotionDetailDTO(dto.getEmotionType(), dto.getPercentage().intValue()))
                 .toList();
     }
 }
